更新系统主页v0.1

+++ src/main/java/cn/edu/asu/bysj/modules/sys/dao/DeptRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DeptRepository extends JpaRepository<Dept,Integer> {
+public interface DeptRepository extends JpaRepository<Dept,Long> {
     // 删除父节点，子节点变成顶级节点（根据实际业务调整）
    // void changeToTop(List<Integer> deptIds);
 }
