+++ src/main/java/cn/edu/asu/bysj/modules/sys/dao/DictRepository.java
@@ -3,5 +3,6 @@ package cn.edu.asu.bysj.modules.sys.dao;
 import cn.edu.asu.bysj.modules.sys.entity.Dict;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DictRepository extends JpaRepository<Dict,Integer>{
+public interface DictRepository extends JpaRepository<Dict,Long>{
+
 }
