+++ src/main/java/cn/edu/asu/bysj/modules/sys/dao/RoleDeptRepository.java
@@ -3,5 +3,5 @@ package cn.edu.asu.bysj.modules.sys.dao;
 import cn.edu.asu.bysj.modules.sys.entity.RoleDept;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleDeptRepository extends JpaRepository<RoleDept,Integer> {
+public interface RoleDeptRepository extends JpaRepository<RoleDept,Long> {
 }
