+++ src/main/java/cn/edu/asu/bysj/modules/sys/dao/RoleMenuRepository.java
@@ -3,5 +3,5 @@ package cn.edu.asu.bysj.modules.sys.dao;
 import cn.edu.asu.bysj.modules.sys.entity.RoleMenu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleMenuRepository extends JpaRepository<RoleMenu,Integer>{
+public interface RoleMenuRepository extends JpaRepository<RoleMenu,Long>{
 }
