+++ src/main/java/cn/edu/asu/bysj/modules/sys/entity/RoleDept.java
@@ -13,32 +13,32 @@ import java.io.Serializable;
 public class RoleDept implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer sys_roledept_id;
-    private Integer role_id;
+    private Long sys_roledept_id;
+    private Long role_id;
 
-    private Integer dept_id;
+    private Long dept_id;
 
-    public Integer getSys_roledept_id() {
+    public Long getSys_roledept_id() {
         return sys_roledept_id;
     }
 
-    public void setSys_roledept_id(Integer sys_roledept_id) {
+    public void setSys_roledept_id(Long sys_roledept_id) {
         this.sys_roledept_id = sys_roledept_id;
     }
 
-    public Integer getRole_id() {
+    public Long getRole_id() {
         return role_id;
     }
 
-    public void setRole_id(Integer role_id) {
+    public void setRole_id(Long role_id) {
         this.role_id = role_id;
     }
 
-    public Integer getDept_id() {
+    public Long getDept_id() {
         return dept_id;
     }
 
-    public void setDept_id(Integer dept_id) {
+    public void setDept_id(Long dept_id) {
         this.dept_id = dept_id;
     }
 }
