+++ src/main/java/cn/edu/asu/bysj/modules/sys/entity/RoleMenu.java
@@ -11,32 +11,32 @@ import java.io.Serializable;
 public class RoleMenu implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private  Integer sys_rolemenu_id;
-    private Integer menu_id;
+    private  Long sys_rolemenu_id;
+    private Long menu_id;
 
-    private Integer role_id;
+    private Long role_id;
 
-    public Integer getSys_rolemenu_id() {
+    public Long getSys_rolemenu_id() {
         return sys_rolemenu_id;
     }
 
-    public void setSys_rolemenu_id(Integer sys_rolemenu_id) {
+    public void setSys_rolemenu_id(Long sys_rolemenu_id) {
         this.sys_rolemenu_id = sys_rolemenu_id;
     }
 
-    public Integer getMenu_id() {
+    public Long getMenu_id() {
         return menu_id;
     }
 
-    public void setMenu_id(Integer menu_id) {
+    public void setMenu_id(Long menu_id) {
         this.menu_id = menu_id;
     }
 
-    public Integer getRole_id() {
+    public Long getRole_id() {
         return role_id;
     }
 
-    public void setRole_id(Integer role_id) {
+    public void setRole_id(Long role_id) {
         this.role_id = role_id;
     }
 }
